+++ src/lib/parser/UcParser.test.ts
@@ -236,6 +236,24 @@ test("parse empty function with local var", () => { parsing(`
     .hasNoErrors();
 });
 
+test("parse empty function with multiple local var in statement", () => { parsing(`
+    function PostBeginPlay(){
+        local int i, j;
+    }
+    `)
+    .hasFunction(0, {
+        name: 'PostBeginPlay',
+        locals: [{
+            type: 'int',
+            name: 'i',
+        },{
+            type: 'int',
+            name: 'j',
+        }]
+    })
+    .hasNoErrors();
+});
+
 test("parse function with empty function call", () => { parsing(`
     function PreBeginPlay(){
         Init();